fix: OneOfSchema now persists the model of each of the child schemas. (#1273)
This allows users to switch between tabs (or radio btn or select) without losing what they typed.

We achieve this in GenericSchema, by adding a key when rendering ObjectSchema. The key is based on... you guessed it: the schema.

In order for this to work we had to remove a recently added useEffect that would re-initialise the model when props.model changed. See details in PR description.